fix: lint

+++ core/rawdb/chain_iterator_test.go
@@ -61,7 +61,7 @@ func TestChainIterator(t *testing.T) {
 			})
 		}
 		txs = append(txs, tx)
-		block = types.NewBlock(&types.Header{Height: uint64(i)}, []*types.Transaction{tx}, nil, newHasher())
+		block = types.NewBlock(&types.Header{Height: i}, []*types.Transaction{tx}, nil, newHasher())
 		WriteBlock(chainDb, block)
 		//WriteCanonicalHash(chainDb, block.Hash(), block.Height())
 		WriteFinalizedHashNumber(chainDb, block.Hash(), block.Height())
@@ -141,7 +141,7 @@ func TestIndexTransactions(t *testing.T) {
 			})
 		}
 		txs = append(txs, tx)
-		block = types.NewBlock(&types.Header{Height: uint64(i)}, []*types.Transaction{tx}, nil, newHasher())
+		block = types.NewBlock(&types.Header{Height: i}, []*types.Transaction{tx}, nil, newHasher())
 		WriteBlock(chainDb, block)
 		//WriteCanonicalHash(chainDb, block.Hash(), block.Height())
 	}
